requirements.txt
asttokens==2.4.1
boto3==1.34.128
botocore==1.34.128
colorama==0.4.6
decorator==5.1.1
executing==2.0.1
ipython==8.25.0
jedi==0.19.1
jmespath==1.0.1
linkify-it-py==2.0.3
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
parso==0.8.4
prompt_toolkit==3.0.47
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rich==13.7.1
s3transfer==0.10.1
six==1.16.0
stack-data==0.6.3
textual==0.69.0
traitlets==5.14.3
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.2.2
wcwidth==0.2.13
